src/FinanceHub.Ingestion/requirements.txt
pandas==2.2.2
pydantic==2.7.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9